streamlit_app/requirements.txt
streamlit
transformers[agents]
sentence-transformers
langchain
ctransformers
llama-cpp-python
auto-gptq
optimum
chromadb-client
# opentelemetry-api
# opentelemetry-sdk
opentelemetry-exporter-otlp
unstructured[all-docs]
python-magic
pdf2image
pdfminer.six
pytesseract
